Methods for investigating if/what application id(s) is/are registered
Added method getRegisteredApplications in NamingService. Method
isApplicationIdRegistered changed to public one.

+Tests

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/branches/ProActive_DataSpaces@12934 28e8926c-6b08-0410-baaa-805c5e19b8d6

+++ src/Tests/unitTests/dataspaces/SpacesDirectoryAbstractBase.java
@@ -24,11 +24,11 @@ import org.objectweb.proactive.extra.dataspaces.exceptions.SpaceAlreadyRegistere
 @Ignore
 public abstract class SpacesDirectoryAbstractBase {
 
-    protected static final int MAIN_APPID = 1;
+    protected static final long MAIN_APPID = 1;
 
-    protected static final int ANOTHER_APPID1 = 0;
+    protected static final long ANOTHER_APPID1 = 0;
 
-    protected static final int ANOTHER_APPID2 = 2;
+    protected static final long ANOTHER_APPID2 = 2;
 
     private SpacesDirectory dir;
